[status] Ensure CO status messages are not empty
Problem:
Marketplace recently encountered a bug that was difficult to diagnose
due to ClusterOperator.Status.Messages being set to an empty strings
in multiple location. This bug would have been easier to diagnose
if the messages were set/unique.

Solution:
Make sure that ClusterOperator.Status.Messages are not empty strings.