[status] Degrade when enabled default opsrc fails
This commit introduces a change so the marketplace operator will report
that it has degraded when one of the enabled default OperatorSources
is unable to reach the succeeded phase.

This change relies on the OperatorHub resource named cluster, which
stores a list of enabled default OperatorSources. If one of these
default enabled OperatorSources fails to reach the succeeded state,
marketplace will report degraded. Marketplace will no longer report
degraded when the problem is resolved or when the OperatorSource is
no longer marked as an enabled default OperatorSource.